requirements.txt
Django~=5.0.3
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.0
django-filter==23.5
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
django-colorfield==0.11.0
channels==4.0.0
django-cors-headers==4.3.1
channels-redis==4.2.0
PyJWT==2.8.0
pillow==10.2.0


